1.fix sql bug 2.添加根据开行日期、委托日期或系统日期生成业务号方法 3.fix 付款申请不能点击应付核销按钮bug

+++ FOS/src/main/java/haitai/fos/sys/entity/table/PSerialRule.java
@@ -18,7 +18,7 @@ public class PSerialRule extends BaseDomain {
 	private String seruRule;
 	private Integer seruSnLength;
 	private String seruUniqSuffix;
-	private Short seruLoopPeriod;
+	private Byte seruLoopPeriod;
 
 	@Id
 	@GeneratedValue(strategy = IDENTITY)
@@ -86,11 +86,11 @@ public class PSerialRule extends BaseDomain {
 	}
 
 	@Column(name = "seru_loop_period", nullable = false)
-	public Short getSeruLoopPeriod() {
+	public Byte getSeruLoopPeriod() {
 		return this.seruLoopPeriod;
 	}
 
-	public void setSeruLoopPeriod(Short seruLoopPeriod) {
+	public void setSeruLoopPeriod(Byte seruLoopPeriod) {
 		this.seruLoopPeriod = seruLoopPeriod;
 	}
 }
